similarity/requirements.txt
# Needed packages/versions
numpy>=1.15.1
python-Levenshtein>=0.12.0
scipy>=1.1.0
gensim>=3.8.0
